emove stitches from sdk
