+++ packages/react-sdk/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./css";
 export * from "./tree";
-export * from "./stitches";
 export * as components from "./components";
 export * as componentsMeta from "./components/meta";
 export * from "./user-props";
