--- packages/react-sdk/src/remix/constants.ts
@@ -1,1 +1,0 @@
-export const CRITICAL_CSS_MARKER = "__CRITICAL_CSS__";
