--- packages/react-sdk/src/stitches/index.ts
@@ -1,2 +1,0 @@
-export * from "./to-css";
-export * from "./stitches";
